fixed install.sh

+++ install.sh
@@ -1,3 +1,9 @@
 ln -sb $(pwd)/vimrc ~/.vimrc
-unlink ~/.vim/after/ftplugin
+if [ ! -d ~/.vim/after ]; then
+    mkdir ~/.vim/after
+fi
+if [ -e ~/.vim/after/ftplugin ]; then
+    echo "error: after/ftplugin exists"
+    exit
+fi
 ln -s $(pwd)/ftplugin ~/.vim/after
